Moved points files
