+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/features/farming/endoffarmnotifer/Range3d.kt
@@ -4,7 +4,7 @@
 //
 package me.partlysanestudios.partlysaneskies.features.farming.endoffarmnotifer
 
-import me.partlysanestudios.partlysaneskies.data.location.Point3d
+import me.partlysanestudios.partlysaneskies.render.points.Point3d
 import kotlin.math.max
 import kotlin.math.min
