+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/features/farming/endoffarmnotifer/RangeHighlight.kt
@@ -27,7 +27,7 @@
 package me.partlysanestudios.partlysaneskies.features.farming.endoffarmnotifer
 
 import me.partlysanestudios.partlysaneskies.PartlySaneSkies
-import me.partlysanestudios.partlysaneskies.data.location.Point3d
+import me.partlysanestudios.partlysaneskies.render.points.Point3d
 import me.partlysanestudios.partlysaneskies.data.skyblockdata.IslandType
 import me.partlysanestudios.partlysaneskies.features.themes.ThemeManager
 import me.partlysanestudios.partlysaneskies.utils.ChatUtils
