+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/render/points/Point2d.kt
@@ -3,7 +3,7 @@
 // See LICENSE for copyright and license notices.
 //
 
-package me.partlysanestudios.partlysaneskies.data.location
+package me.partlysanestudios.partlysaneskies.render.points
 
 open class Point2d(val x: Double, val y: Double) {
     fun getPointX(): Double {
