+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/render/points/Point3d.kt
@@ -2,7 +2,7 @@
 // Written by Su386 and J10a1n15.
 // See LICENSE for copyright and license notices.
 //
-package me.partlysanestudios.partlysaneskies.data.location
+package me.partlysanestudios.partlysaneskies.render.points
 
 open class Point3d(x: Double, y: Double, val z: Double) : Point2d(x, y) {
     fun getPointZ(): Double {
