+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/utils/MathUtils.kt
@@ -7,8 +7,8 @@
 package me.partlysanestudios.partlysaneskies.utils
 
 import me.partlysanestudios.partlysaneskies.PartlySaneSkies
-import me.partlysanestudios.partlysaneskies.data.location.Point2d
-import me.partlysanestudios.partlysaneskies.data.location.Point3d
+import me.partlysanestudios.partlysaneskies.render.points.Point2d
+import me.partlysanestudios.partlysaneskies.render.points.Point3d
 import java.util.concurrent.ThreadLocalRandom
 import kotlin.math.pow
 import kotlin.math.sqrt
